Added Jape file
Added a WIP Jape file for proofs. As of right now, the file only focuses on 3 proofs and doesn't have any relevant conclusions. Would greatly appreciate feedback/guidance on where we can go with this (what sort of conclusions to make).

As a side note, these formulas do not match those of our PowerPoint as Jape does not accept any letter as a variable, and does not accept numerical subscripts. If anyone can figure out how to do these things in Jape, that would be greatly appreciated!