do the constraints properly
before, we applied constraints based on whether we detect a certain
piece on a square. this version applies the constraints to every square,
which is what should have been done from the beginning. set_board was
reverted back to the original state because otherwise there wouldn't be
a constraint that prevents the model from placing its own pieces.